Backend2/requirements.txt
fastapi
uvicorn
sqlalchemy
pymysql
python-dotenv
python-jose[cryptography]
passlib[bcrypt]
python-multipart

